gen: Reset the makefile to fix the gp fault
I got a General protection fault on irq call. After resetting the
Makefile to how it was before I touched it, it all seemed to work
again. Lesson learned. Don't touch those Makefiles. All you'll do
is get hurt.

Did some other stuff, I don't know exactly what, but I do know it
didn't hurt the stability. The contrary. I beleive I optimised one
data structure and fixed the assembly of using it to what it
should be.

+++ src/include/arch/x86/task.h
@@ -40,10 +40,10 @@ typedef struct general_registers
 
 struct isr_regs
 {
-        uint16_t ds;
-        unsigned long edi, esi, ebp, esp, ebx, edx, ecx, eax;
-        unsigned long funcPtr, errCode;
-        unsigned long eip, cs, eflags, procesp, ss;
+        uint32_t ds;
+        uint32_t edi, esi, ebp, esp, ebx, edx, ecx, eax;
+        uint32_t funcPtr, errCode;
+        uint32_t eip, cs, eflags, procesp, ss;
 } __attribute__((packed));
 typedef struct isr_regs isrVal_t;
